Point_set:  move IO code to Stream_support
