+++ Point_set_3/test/Point_set_3/point_set_test_join.cpp
@@ -1,7 +1,6 @@
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 
 #include <CGAL/Point_set_3.h>
-#include <CGAL/IO/write_xyz_points.h>
 #include <CGAL/grid_simplify_point_set.h>
 
 #include <fstream>
