+++ Stream_support/include/CGAL/IO/OFF/read_off_points.h
@@ -1,18 +1,21 @@
-// Copyright (c) 2007-09  INRIA Sophia-Antipolis (France).
-// All rights reserved.
+// Copyright (c) 1997
+// Utrecht University (The Netherlands),
+// ETH Zurich (Switzerland),
+// INRIA Sophia-Antipolis (France),
+// Max-Planck-Institute Saarbruecken (Germany),
+// and Tel-Aviv University (Israel).  All rights reserved.
 //
-// This file is part of CGAL (www.cgal.org).
+// This file is part of CGAL (www.cgal.org);
 //
 // $URL$
 // $Id$
-// SPDX-License-Identifier: GPL-3.0-or-later OR LicenseRef-Commercial
+// SPDX-License-Identifier: LGPL-3.0-or-later OR LicenseRef-Commercial
 //
 // Author(s) : Pierre Alliez and Laurent Saboret
 
-#ifndef CGAL_POINT_SET_PROCESSING_READ_OFF_POINTS_H
-#define CGAL_POINT_SET_PROCESSING_READ_OFF_POINTS_H
+#ifndef CGAL_IO_OFF_READ_OFF_POINTS_H
+#define CGAL_IO_OFF_READ_OFF_POINTS_H
 
-#include <CGAL/license/Point_set_processing_3.h>
 
 #include <CGAL/IO/io.h>
 #include <CGAL/property_map.h>
@@ -273,4 +276,4 @@ bool read_OFF(const std::string& fname, OutputIterator output, const CGAL_NP_CLA
 
 } // namespace CGAL
 
-#endif // CGAL_POINT_SET_PROCESSING_READ_OFF_POINTS_H
+#endif // CGAL_IO_OFF_READ_OFF_POINTS_H
