+++ Stream_support/include/CGAL/IO/PLY/read_ply_points.h
@@ -1,18 +1,16 @@
-// Copyright (c) 2015  Geometry Factory
-// All rights reserved.
+// Copyright (c) 2017  GeometryFactory
 //
-// This file is part of CGAL (www.cgal.org).
+// This file is part of CGAL (www.cgal.org);
 //
 // $URL$
 // $Id$
-// SPDX-License-Identifier: GPL-3.0-or-later OR LicenseRef-Commercial
+// SPDX-License-Identifier: LGPL-3.0-or-later OR LicenseRef-Commercial
 //
 // Author(s) : Simon Giraudot
 
-#ifndef CGAL_POINT_SET_PROCESSING_READ_PLY_POINTS_H
-#define CGAL_POINT_SET_PROCESSING_READ_PLY_POINTS_H
+#ifndef CGAL_IO_PLY_READ_PLY_POINTS_H
+#define CGAL_IO_PLY_READ_PLY_POINTS_H
 
-#include <CGAL/license/Point_set_processing_3.h>
 
 #include <CGAL/config.h>
 
@@ -364,4 +362,4 @@ bool read_PLY(const std::string& fname, OutputIterator output, const CGAL_NP_CLA
 #undef TRY_TO_GENERATE_SIZED_FACE_PROPERTY
 #undef TRY_TO_GENERATE_FACE_PROPERTY
 
-#endif // CGAL_POINT_SET_PROCESSING_READ_PLY_POINTS_H
+#endif // CGAL_IO_PLY_READ_PLY_POINTS_H
