+++ Stream_support/include/CGAL/IO/PLY/write_ply_points.h
@@ -1,18 +1,20 @@
-// Copyright (c) 2015  Geometry Factory
-// All rights reserved.
+// Copyright (c) 1997
+// Utrecht University (The Netherlands),
+// ETH Zurich (Switzerland),
+// INRIA Sophia-Antipolis (France),
+// Max-Planck-Institute Saarbruecken (Germany),
+// and Tel-Aviv University (Israel).  All rights reserved.
 //
-// This file is part of CGAL (www.cgal.org).
+// This file is part of CGAL (www.cgal.org);
 //
 // $URL$
 // $Id$
-// SPDX-License-Identifier: GPL-3.0-or-later OR LicenseRef-Commercial
+// SPDX-License-Identifier: LGPL-3.0-or-later OR LicenseRef-Commercial
 //
 // Author(s) : Simon Giraudot
 
-#ifndef CGAL_POINT_SET_PROCESSING_WRITE_PLY_POINTS_H
-#define CGAL_POINT_SET_PROCESSING_WRITE_PLY_POINTS_H
-
-#include <CGAL/license/Point_set_processing_3.h>
+#ifndef CGAL_IO_PLY_WRITE_PLY_POINTS_H
+#define CGAL_IO_PLY_WRITE_PLY_POINTS_H
 
 #include <CGAL/IO/helpers.h>
 #include <CGAL/IO/PLY.h>
@@ -300,4 +302,4 @@ bool write_PLY(const std::string& filename,
 
 } // namespace CGAL
 
-#endif // CGAL_POINT_SET_PROCESSING_WRITE_PLY_POINTS_H
+#endif // CGAL_IO_PLY_WRITE_PLY_POINTS_H
